+++ src/main/java/subway/config/FareConfiguration.java
@@ -2,8 +2,8 @@ package subway.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import subway.domain.BasicFarePolicy;
-import subway.domain.FarePolicy;
+import subway.domain.fare.BasicFarePolicy;
+import subway.domain.fare.FarePolicy;
 
 @Configuration
 public class FareConfiguration {
