+++ src/main/java/subway/domain/path/Path.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.path;
 
 import subway.domain.section.Distance;
